Small updates: no AI changes
removed useless variables and renamed ship counter in MyBot.cpp
Location now uses Vector2 for position
bot now times how long each turn was
new Vector2 class

+++ Halite 2 Bot/hlt/constants.hpp
@@ -54,6 +54,6 @@ namespace hlt {
          * Used in Location::get_closest_point()
          * Minimum distance specified from the object's outer radius.
          */
-        constexpr int MIN_DISTANCE_FOR_CLOSEST_POINT = 3;
+        constexpr double MIN_DISTANCE_FOR_CLOSEST_POINT = 3;
     }
 }
